requirements.txt
click==8.1.7
keras==3.3.3
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
scikit_learn==1.4.2
setuptools==68.0.0
tensorflow==2.16.1
dvc[all]==3.5.0